refactor: LadderLine domain.line로 패키지 이동
