+++ src/main/java/ladder/Application.java
@@ -1,7 +1,7 @@
 package ladder;
 
 import ladder.application.LadderGame;
-import ladder.domain.LadderLinePainter;
+import ladder.domain.line.LadderLinePainter;
 import ladder.domain.PositionLogFactory;
 import ladder.view.LadderGameController;
