+++ src/main/java/ladder/domain/line/LadderLines.java
@@ -1,4 +1,4 @@
-package ladder.domain;
+package ladder.domain.line;
 
 import java.util.Collections;
 import java.util.List;
