+++ src/test/java/ladder/PositionLogsTest.java
@@ -1,6 +1,8 @@
 package ladder;
 
 import ladder.domain.*;
+import ladder.domain.line.LadderLine;
+import ladder.domain.line.LadderLines;
 import ladder.domain.point.DrawingPointStrategy;
 import ladder.domain.point.LadderLinePoint;
 import org.junit.jupiter.api.DisplayName;
