+++ src/test/java/ladder/domain/line/LadderLinePainterTest.java
@@ -1,6 +1,7 @@
-package ladder;
+package ladder.domain.line;
 
-import ladder.domain.*;
+import ladder.domain.Player;
+import ladder.domain.Players;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
